Create holiday.js to generate holidays by year
This module is not yet exported or used by index.js (still to do)
It creates an object that contains the names of each holiday in the year
including variable ones like Easter, indexed by month then date.
Regional anniversary dates are not included; some regional anniversaries
are very hard to calculate; implementing them will be difficult.
Mondayisation is not considered and will be handled separately.